bootcamp/week-11/2-practicas/02-jwt-tokens/starter/pyproject.toml
[project]
name = "jwt-tokens-practice"
version = "0.1.0"
description = "Práctica de JWT tokens"
requires-python = ">=3.14"
dependencies = [
    "python-jose[cryptography]>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.0",
    "pytest-asyncio>=0.24.0",
    "freezegun>=1.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
